// ombi-api/model/ombi-store-entities-ombi-user.ts
export interface OmbiStoreEntitiesOmbiUser {
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    alias?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    userType?: OmbiStoreEntitiesOmbiUserUserTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    providerUserId?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    lastLoggedIn?: Date;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    embyConnectUserId?: string;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    movieRequestLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    episodeRequestLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    musicRequestLimit?: number;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    userAccessToken?: string;
    /**
     * 
     * @type {Array<OmbiStoreEntitiesNotificationUserId>}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    notificationUserIds?: Array<OmbiStoreEntitiesNotificationUserId>;
    /**
     * 
     * @type {Array<OmbiStoreEntitiesUserNotificationPreferences>}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    userNotificationPreferences?: Array<OmbiStoreEntitiesUserNotificationPreferences>;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    isEmbyConnect?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    userAlias?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    emailLogin?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    isSystemUser?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    userName?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    normalizedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    normalizedEmail?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    emailConfirmed?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    phoneNumber?: string;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    phoneNumberConfirmed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    twoFactorEnabled?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    lockoutEnd?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    lockoutEnabled?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiStoreEntitiesOmbiUser
     */
    accessFailedCount?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum OmbiStoreEntitiesOmbiUserUserTypeEnum {
    SystemUser = 'SystemUser',
    LocalUser = 'LocalUser',
    PlexUser = 'PlexUser',
    EmbyUser = 'EmbyUser'
}