// ombi-api/model/ombi-core-models-uiuser-view-model.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * Ombi Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { OmbiCoreModelsRequestQuotaCountModel } from './ombi-core-models-request-quota-count-model';
import { OmbiCoreModelsUIClaimCheckboxes } from './ombi-core-models-uiclaim-checkboxes';
import { OmbiStoreEntitiesUserQualityProfiles } from './ombi-store-entities-user-quality-profiles';

/**
 * 
 * @export
 * @interface OmbiCoreModelsUIUserViewModel
 */
export interface OmbiCoreModelsUIUserViewModel {
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    userName?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    alias?: string;
    /**
     * 
     * @type {Array<OmbiCoreModelsUIClaimCheckboxes>}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    claims?: Array<OmbiCoreModelsUIClaimCheckboxes>;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    emailAddress?: string;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    password?: string;
    /**
     * 
     * @type {Date}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    lastLoggedIn?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    hasLoggedIn?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    userType?: OmbiCoreModelsUIUserViewModelUserTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    movieRequestLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    episodeRequestLimit?: number;
    /**
     * 
     * @type {OmbiCoreModelsRequestQuotaCountModel}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    episodeRequestQuota?: OmbiCoreModelsRequestQuotaCountModel;
    /**
     * 
     * @type {OmbiCoreModelsRequestQuotaCountModel}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    movieRequestQuota?: OmbiCoreModelsRequestQuotaCountModel;
    /**
     * 
     * @type {OmbiCoreModelsRequestQuotaCountModel}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    musicRequestQuota?: OmbiCoreModelsRequestQuotaCountModel;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    musicRequestLimit?: number;
    /**
     * 
     * @type {OmbiStoreEntitiesUserQualityProfiles}
     * @memberof OmbiCoreModelsUIUserViewModel
     */
    userQualityProfiles?: OmbiStoreEntitiesUserQualityProfiles;
}

/**
    * @export
    * @enum {string}
    */
export enum OmbiCoreModelsUIUserViewModelUserTypeEnum {
    LocalUser = 'LocalUser',
    PlexUser = 'PlexUser',
    EmbyUser = 'EmbyUser'
}