// ombi-api/model/ombi-core-models-request-quota-count-model.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * Ombi Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface OmbiCoreModelsRequestQuotaCountModel
 */
export interface OmbiCoreModelsRequestQuotaCountModel {
    /**
     * 
     * @type {boolean}
     * @memberof OmbiCoreModelsRequestQuotaCountModel
     */
    hasLimit?: boolean;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsRequestQuotaCountModel
     */
    limit?: number;
    /**
     * 
     * @type {number}
     * @memberof OmbiCoreModelsRequestQuotaCountModel
     */
    remaining?: number;
    /**
     * 
     * @type {Date}
     * @memberof OmbiCoreModelsRequestQuotaCountModel
     */
    nextRequest?: Date;
}